server/requirements.txt
openenv-core>=0.2.3
fastapi
uvicorn[standard]
pydantic
httpx
websockets
openai
